--- applications/stable-diffusion/dreambooth/dreambooth.sh
@@ -1,12 +1,0 @@
-python train_dreambooth.py \
-    --pretrained_model_name_or_path="CompVis/stable-diffusion-v1-4" \
-    --instance_data_dir="/home/lclcq/ColossalAI/applications/stable-diffusion/dreambooth/dog" \
-    --output_dir="./weight_output" \
-    --instance_prompt="a photo of a dog" \
-    --resolution=512 \
-    --train_batch_size=1 \
-    --gradient_accumulation_steps=1 \
-    --learning_rate=5e-6 \
-    --lr_scheduler="constant" \
-    --lr_warmup_steps=0 \
-    --num_class_images=200 \
