--- applications/stable-diffusion/dreambooth/requirements.txt
@@ -1,7 +1,0 @@
-diffusers>==0.5.0
-accelerate
-torchvision
-transformers>=4.21.0
-ftfy
-tensorboard
-modelcards
