+++ applications/stable-diffusion/text_img2img/download_dataset_dreambooth.py
@@ -6,3 +6,4 @@ snapshot_download(
     local_dir=local_dir, repo_type="dataset",
     ignore_patterns=".gitattributes",
 )
+
