+++ applications/stable-diffusion/text_img2img/scripts_run/inference.sh
