+++ applications/stable-diffusion/text_img2img/scripts_run/trainer_colossalai_dreambooth.sh
@@ -2,9 +2,9 @@ HF_DATASETS_OFFLINE=1
 TRANSFORMERS_OFFLINE=1
 DIFFUSERS_OFFLINE=1
 
-torchrun --nproc_per_node 4 --standalone train_dreambooth_colossalai.py \
+torchrun --nproc_per_node 4 --standalone stable_diffusion_colossalai_trainer.py \
   --pretrained_model_name_or_path="CompVis/stable-diffusion-v1-4"  \
-  --instance_data_dir="/home/lclcq/ColossalAI/applications/stable-diffusion/dreambooth/dog" \
+  --instance_data_dir="/home/lclcq/ColossalAI/applications/stable-diffusion/text_img2img/dog" \
   --output_dir="./weight_output" \
   --instance_prompt="a picture of a dog" \
   --resolution=512 \
@@ -14,5 +14,6 @@ torchrun --nproc_per_node 4 --standalone train_dreambooth_colossalai.py \
   --lr_scheduler="constant" \
   --lr_warmup_steps=0 \
   --num_class_images=200 \
-  --test_run=True \
-  --placement="auto" \
+  --placement="cuda" \
+  --task_type="dreambooth" 
+
