+++ applications/stable-diffusion/text_img2img/scripts_run/trainer_no_colossalai_image_to_image.sh
